% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dashboard.R
\name{call_dashboard}
\alias{call_dashboard}
\title{Visualizing validation results according to four steps, namely, set-selection, estimation, diagnosis, and validation}
\usage{
call_dashboard(source.obj = NULL, target.obj = NULL, source.obj.rep = NULL)
}
\arguments{
\item{source.obj}{an instantiated object of class \code{TEstimator}. The estimates of conditional average treatment effects are compared to those from \code{target.obj}.}

\item{target.obj}{an instantiated object of class \code{TEstimator}. The estimates of conditional average treatment effects are regarded as unbiased of truth.}

\item{source.obj.rep}{an instantiated object of class \code{SEstimator}. The estimates of conditional average treatment effects are compared to those from \code{target.obj}.}
}
\value{
an interactive interface visualizing results of four steps
}
\description{
Visualizing validation results according to four steps, namely, set-selection, estimation, diagnosis, and validation
}
