% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_base.R
\name{DR_base}
\alias{DR_base}
\title{R6 class: Doubly robust estimator base class}
\description{
A base R6 class for doubly robust estimator of average treatment effect that implements comment methods.
}
\section{Super class}{
\code{\link[RCTrep:Estimator]{RCTrep::Estimator}} -> \code{DR_base}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DR_base$new()}}
\item \href{#method-clone}{\code{DR_base$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="RCTrep" data-topic="Estimator" data-id="RCTrep">}\href{../../RCTrep/html/Estimator.html#method-RCTrep}{\code{RCTrep::Estimator$RCTrep()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="RCTrep" data-topic="Estimator" data-id="get_CATE">}\href{../../RCTrep/html/Estimator.html#method-get_CATE}{\code{RCTrep::Estimator$get_CATE()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="RCTrep" data-topic="Estimator" data-id="plot_CATE">}\href{../../RCTrep/html/Estimator.html#method-plot_CATE}{\code{RCTrep::Estimator$plot_CATE()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR_base$new(df, vars_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DR_base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
