% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_base.R
\name{Estimator}
\alias{Estimator}
\title{R6 class: Estimator base class}
\description{
A base R6 class for estimator of average treatment effect that implements the common methods, such as \code{\link{RCTrep}}, \code{\link{get_CATE}}, \code{\link{plot_CATE}}, inheritted by \code{\link{G_computation_base}}, \code{\link{IPW_base}}, and \code{\link{DR_base}} class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Estimator$new()}}
\item \href{#method-RCTrep}{\code{Estimator$RCTrep()}}
\item \href{#method-get_CATE}{\code{Estimator$get_CATE()}}
\item \href{#method-plot_CATE}{\code{Estimator$plot_CATE()}}
\item \href{#method-clone}{\code{Estimator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Estimator} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimator$new(df, vars_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{A data frame containing variables in \code{vars_name}}

\item{\code{vars_name}}{vars_name A list containing four vectors \code{confounders_internal}, \code{confounders_external}, \code{treatment_name}, and \code{outcome_name}. \code{confounders_internal} is a character vector containing the adjustment variables, which, along with \code{Estimator} and the corresponding \code{outcome_method} or \code{treatment_method} to correct for confounding; \code{confounders_external} is a character vector containing variables for weighting as to generalize estimates from \code{source.data} to \code{target.data}; \code{outcome_name} is a character vector of length one containing the variable name of outcome; \code{treatment_name} is a character vector of length one containing the variable name of treatment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RCTrep"></a>}}
\if{latex}{\out{\hypertarget{method-RCTrep}{}}}
\subsection{Method \code{RCTrep()}}{
Replicating the average treatment effect of \code{target.obj}. If \code{stratification} is specified, then replicating the conditional average treatment effect stratified by \code{stratification} and \code{stratification_joint} by weighting based on the residual variables, namely, variables that are specified in \code{confounders_external_name} while not in \code{stratification}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimator$RCTrep(
  target.obj,
  weighting_estimator,
  weighting_method,
  stratification,
  stratification_joint
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.obj}}{An object of class \code{Estimator} or list.}

\item{\code{weighting_estimator}}{A string specifying a weighting estimator for generalizing/transporting the estimates to \code{target.obj}. The allowed estimators are: \code{"balancing"}, and \code{"modeling"}.}

\item{\code{weighting_method}}{A string specifying which model for selection to use. Possible values are found using \code{names(getModelInfo())}. See \url{http://topepo.github.io/caret/train-models-by-tag.html}.}

\item{\code{stratification}}{An optional string vector containing variables to define subgroup. If \code{!is.NULL(stratification)}, \code{source.obj} will compute both weighted and unweighted conditional average treatment effect based on these variables, \code{target.obj} will calculate the conditional average treatment effect based on these variables.}

\item{\code{stratification_joint}}{An optional logical defining the subgroup based on joint distribution of variables or univariate distribution in \code{stratification} when \code{stratification} is specified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_CATE"></a>}}
\if{latex}{\out{\hypertarget{method-get_CATE}{}}}
\subsection{Method \code{get_CATE()}}{
Get conditional average treatment effect of subgroups defined by \code{stratification} and \code{stratification_joint}. If \code{stratification_joint=FALSE}, then the method return conditional average treatment effect of subgroups stratified by each of variables in \code{stratification}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimator$get_CATE(stratification, stratification_joint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stratification}}{An string vector containing variables to define subgroup.}

\item{\code{stratification_joint}}{An logical defining the subgroup based on joint distribution of variables or univariate distribution in \code{stratification}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame. If \code{stratification_joint=TRUE}, then the method returns a data frame with N rows and J columns, where N represents the number of subgroups, and J is equal to the sum of number of variables in \code{stratification} and 3 (three additional columns with name \code{cate}, \code{se}, and \code{size}, representing the estimated conditional average treatment effect of this subgroup, standard error of the estimate, and the sample size of the subgroup). If \code{stratification_joint=FALSE}, then the method returns a data frame with N rows and 5 columns, where N represents the number of subgroups stratified by each variable in \code{stratification} and 5 columns with name \code{name}, \code{value}, \code{cate}, \code{se}, and \code{size}, representing the name of a variable used to stratify the population, a level of the variable, the estimated conditional average treatment effect of this subgroup, standard error of the estimate, and the sample size of the subgroup).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_CATE"></a>}}
\if{latex}{\out{\hypertarget{method-plot_CATE}{}}}
\subsection{Method \code{plot_CATE()}}{
Plot the forest plot of conditional average treatment effect of subgroups defined by \code{stratification} and \code{stratification_joint}. The method first call public method \code{get_CATE(stratification,stratification_joint)}, then plot the results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimator$plot_CATE(stratification, stratification_joint = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stratification}}{An string vector containing variables to define subgroup.}

\item{\code{stratification_joint}}{An logical defining the subgroup based on joint distribution of variables or univariate distribution in \code{stratification}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot containing a forest plot and a table with numeric results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Estimator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
