% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fusion.R
\name{Fusion}
\alias{Fusion}
\title{Validation of estimates of conditional average treatment effects in objects of class \code{TEstimator} and \code{SEstimator}.}
\description{
Validation of estimates of conditional average treatment effects in objects of class \code{TEstimator} and \code{SEstimator}.

Validation of estimates of conditional average treatment effects in objects of class \code{TEstimator} and \code{SEstimator}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Fusion$new()}}
\item \href{#method-plot}{\code{Fusion$plot()}}
\item \href{#method-print}{\code{Fusion$print()}}
\item \href{#method-evaluate}{\code{Fusion$evaluate()}}
\item \href{#method-clone}{\code{Fusion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fusion$new(..., stratification = NULL, stratification_joint = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{objects of class \code{TEstimator} and \code{SEstimator}.}

\item{\code{stratification}}{a character vector specifying variables. The variables are used to select sub-populations individually or in combination depending on \code{stratification_joint}. Default value is NULL.}

\item{\code{stratification_joint}}{a logical indicating if subpopulations are selected based on levels of individual variable in \code{stratification} or levels of combined variables in \code{stratifiation}. Default value is NULL.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fusion$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fusion$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fusion$evaluate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Fusion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
