% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{Plot_estimates}
\alias{Plot_estimates}
\title{plot average treatment effect and possible conditional average treatment}
\usage{
Plot_estimates(source.obj, target.obj)
}
\arguments{
\item{source.obj}{An object of class \code{\link{Estimator}} resulting from source data}

\item{target.obj}{An object of class \code{\link{Estimator}} resulting from target data}
}
\value{
a plot with a forest plot and a table with numeric results
}
\description{
plot average treatment effect and possible conditional average treatment
}
