% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quasar.agg}
\alias{quasar.agg}
\title{Aggregated data derived from paper of QUASAR trial}
\format{
An object of class \code{list} of length 5.
}
\usage{
quasar.agg
}
\description{
Aggregated data derived from paper of QUASAR trial
}
\keyword{datasets}
