% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quasar.synthetic}
\alias{quasar.synthetic}
\title{A synthetic QUASAR trial dataset, where outcome is a binary variable, treatment is a binary variable.}
\format{
## 'quasar.synthetic'
A data frame with 5934 rows and 3 variables:
\describe{
\item{Stage2}{binary variable, 1 indicating stage 2 and 0 indicating stage 3}
\item{male}{binary variable, 1 indicating male and 0 indicating female}
\item{age}{categorical variable, 1 indicating [23,50], 2 indicating [50,59], 3 indicating [60,69], 4 indicating [70,86]}
}
}
\usage{
quasar.synthetic
}
\description{
A synthetic QUASAR trial dataset, where outcome is a binary variable, treatment is a binary variable.
}
\keyword{datasets}
