% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTrep.R
\name{RCTREP}
\alias{RCTREP}
\title{Replicate treatment effect estimates obtained from a randomized control trial using real world data}
\usage{
RCTREP(
  Estimator = "G_computation",
  weighting_estimator = "Balancing",
  source.data = source.data,
  target.data = target.data,
  vars_name,
  outcome_method = "glm",
  treatment_method = "glm",
  weighting_model = "glm",
  outcome_formula = NULL,
  treatment_formula = NULL,
  selection_formula = NULL,
  stratification = NULL,
  stratification_joint = FALSE,
  strata_cut = NULL,
  two_models = NULL,
  ...
)
}
\arguments{
\item{Estimator}{a character specifying an estimator for average treatment effect. The allowed estimators for \code{Estimator} are: \code{"G_computation"}, \code{"IPW"}, and \code{"DR"}. The corresponding object will be created by the wrapper function \code{\link{Estimate()}}.}
}
\value{

}
\description{
The function \code{RCTREP} is used to replicate the estimate of treatment effect from a target randomized control trial based on non-experimental observational data. This function estimate the treatment effect of observational data to ensure the internal validity of the estimates within the study population (namely, the observational data) and weight the resulting estimates to the target population (namely, the RCT) to enable external validity. The function currently implement the following types of estimators of treatment effect: G_computation, inverse propensity score weighting (IPW), and augmented propensity score weighting. The function implement the following two types of weighting estimators to generalize the resulting estimates of treatment effect from observational data to the target randomized control trial population: exact matching weights, and selection score weights. Since the target population is the RCT, weights for each individual in observational data is \eqn{p/(1-p)}.
}
\details{
An R6 object for both studies are constructed by a wrapper function \code{Estimate()} with user's input of data and estimator for treatment effect. Then \code{Estimate()} return initialized objects \code{source.obj} and \code{target.obj}. \code{source.obj} replicates the target RCT estimate via the class method \code{RCTrep()} with input of target object \code{target.obj} and weighting method.
}
