% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{source.data}
\alias{source.data}
\title{Dutch Colon cancer registry data}
\format{
A data frame with 2000 rows and 9 variables
\describe{
\item{vitstat}{1=death, 0=survival}
\item{Stage2}{pathological stage, 1=stage2, 0=stage3}
}
}
\usage{
source.data
}
\description{
A dataset containing synthetic dutch colon cancer registry data and other biomarkers
}
\keyword{datasets}
