% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{summary.Estimator}
\alias{summary.Estimator}
\title{Summary average treatment effect and possible conditional average treatment}
\usage{
summary.Estimator(target.obj, source.obj)
}
\arguments{
\item{target.obj}{An object of class \code{\link{Estimator}} resulting from target data}

\item{source.obj}{An object of class \code{\link{Estimator}} resulting from source data, the field estimates$CATE,estimates$CATE_weighted should have the same dimension and the same order as target.obj, if no, need preprocess.}
}
\value{
a plot with a forest plot and a table with numeric results
}
\description{
Summary average treatment effect and possible conditional average treatment
}
