% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_computation.R
\name{G_computation}
\alias{G_computation}
\title{R6 class: G_computation base class}
\description{
A base R6 class for G_computation estimator for average treatment effect
}
\section{Super class}{
\code{\link[RCTrep:Estimator]{RCTrep::Estimator}} -> \code{G_computation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{G_computation$new()}}
\item \href{#method-residual_check}{\code{G_computation$residual_check()}}
\item \href{#method-summary}{\code{G_computation$summary()}}
\item \href{#method-clone}{\code{G_computation$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="RCTrep" data-topic="Estimator" data-id="EstimateRep">}\href{../../RCTrep/html/Estimator.html#method-EstimateRep}{\code{RCTrep::Estimator$EstimateRep()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="RCTrep" data-topic="Estimator" data-id="get_CATE">}\href{../../RCTrep/html/Estimator.html#method-get_CATE}{\code{RCTrep::Estimator$get_CATE()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="RCTrep" data-topic="Estimator" data-id="plot_CATE">}\href{../../RCTrep/html/Estimator.html#method-plot_CATE}{\code{RCTrep::Estimator$plot_CATE()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{G_computation$new(
  df,
  vars_name,
  name,
  gc.method,
  gc.formula,
  var_approach = "Bias_adjusted",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-residual_check"></a>}}
\if{latex}{\out{\hypertarget{method-residual_check}{}}}
\subsection{Method \code{residual_check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{G_computation$residual_check(stratification)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{G_computation$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{G_computation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
