% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quasar.obj}
\alias{quasar.obj}
\title{An object of class TEstimator_Synthetic using quasar.synthetic}
\format{
An object of class \code{TEstimator_Synthetic} (inherits from \code{TEstimator}, \code{R6}) of length 15.
}
\usage{
quasar.obj
}
\description{
An object of class TEstimator_Synthetic using quasar.synthetic
}
\keyword{datasets}
