% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTrep.R
\name{RCTREP}
\alias{RCTREP}
\title{Replicate treatment effect estimates obtained from a randomized control trial using real world data}
\usage{
RCTREP(
  Estimator = "G_computation",
  weighting_estimator = "Balancing",
  source.data = source.data,
  target.data = target.data,
  vars_name,
  outcome_method = "glm",
  treatment_method = "glm",
  weighting_model = "glm",
  outcome_formula = NULL,
  treatment_formula = NULL,
  selection_formula = NULL,
  stratification = NULL,
  stratification_joint = FALSE,
  strata_cut = NULL,
  two_models = NULL,
  ...
)
}
\arguments{
\item{Estimator}{a character specifying an estimator for average treatment effect. The allowed estimators for \code{Estimator} are: \code{"G_computation"}, \code{"IPW"}, and \code{"DR"}. The corresponding object will be created by the wrapper function \code{\link{Estimate()}}.}

\item{weighting_estimator}{a character specifying a weighting estimator for generalizing/transporting the estimates of \code{source.obj} (initiated using RWD) to \code{target.obj} (initiated using RCT) as to enable replication/comparison between source study (RWD) and target study(RCT). The allowed estimators are: \code{"balancing"}, and \code{"modeling"}. \code{"balancing"} estimator use exact matching to compute the weight; \code{"modeling"} estimator model the probability of being selected to the target RCT (assuming combing target RCT and the source RWD as a population).}

\item{source.data}{a data frame containing variables named in \code{vars_name} and possible other variables. If not found in \code{source.data}, the function will stop and throw error.}

\item{target.data}{a data frame or a list}
}
\value{

}
\description{
The function \code{RCTREP} is used to replicate the estimate of treatment effect from a target randomized control trial based on real-world data (RWD). This function estimate the treatment effect of RWD to ensure the internal validity of the estimates within the study population (namely, the observational data) and weight the resulting estimates to the target population (namely, the RCT) to enable external validity. The function currently implement the following types of estimators of treatment effect: G_computation, inverse propensity score weighting (IPW), and augmented propensity score weighting. The function implement the following two types of weighting estimators to generalize the resulting estimates of treatment effect from RWD to the target RCT: exact matching weights, and selection score weights. Since we regard the sample in the RCT as the target population, weights for each individual in RWD is \eqn{p/(1-p)} so that the weighted population of RWD is representative to the target population.
}
\details{
An R6 object for both studies are constructed by a wrapper function \code{Estimate()} with user's input of data and estimator for treatment effect. Then \code{Estimate()} return initialized objects \code{source.obj} and \code{target.obj}. \code{source.obj} replicates the target RCT estimate via the class method \code{RCTrep()} with input of target object \code{target.obj} and weighting method.
}
