% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEstimator_wrapper.R
\name{SEstimator_wrapper}
\alias{SEstimator_wrapper}
\title{R6 class which estimates the weighted conditional average treatment effects in \code{source.obj} based on input objects \code{source.obj} and \code{target.obj} of class \code{\link{TEstimator}}.}
\usage{
SEstimator_wrapper(
  Estimator,
  target.obj,
  source.obj,
  confounders_sampling_name,
  method = "glm",
  sampling_formula = NULL,
  ...
)
}
\arguments{
\item{Estimator}{a character specifying an estimator for weight. The allowed estimators are \code{Exact}, \code{ISW}, and \code{Subclass}.}

\item{target.obj, source.obj}{an instantiated object of class \code{TEstimator}.}

\item{confounders_sampling_name}{a character vector specifying the names of variables in \code{data} of \code{source.obj} and \code{target.obj}. Weights are estimated based on the variables.}

\item{method}{an optional character specifying a model for estimating sampling probability once \code{Estimator='ISW'} or \code{Estimator='Subclass'}.}

\item{sampling_formula}{an object of class \code{formula} specifying a model specification for sampling probability. Default value is NULL.}

\item{...}{an optional argument specifying training and tuning for a model of sampling probability. See \url{https://topepo.github.io/caret/model-training-and-tuning.html} for details.}
}
\description{
R6 class which estimates the weighted conditional average treatment effects in \code{source.obj} based on input objects \code{source.obj} and \code{target.obj} of class \code{\link{TEstimator}}.
}
