% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEstimate_wrapper.R
\name{TEstimator_wrapper}
\alias{TEstimator_wrapper}
\title{Estimator for average treatment effect wrapper function}
\usage{
TEstimator_wrapper(
  Estimator,
  data,
  vars_name,
  name = "",
  outcome_method = "glm",
  treatment_method = "glm",
  two_models = FALSE,
  outcome_formula = NULL,
  treatment_formula = NULL,
  data.public = TRUE,
  isTrial = FALSE,
  ...
)
}
\arguments{
\item{Estimator}{A character specifying an estimator for average treatment effect. The allowed estimators for \code{Estimator} are: \code{"G_computation"}, \code{"IPW"}, and \code{"DR"}. The corresponding object will be created by the wrapper function \code{\link{Estimate()}}. The default is \code{"G_computation"}, which, along with \code{outcome_method="glm"} model the potential outcomes.}

\item{data}{A data frame containing variables in \code{vars_name}.}

\item{vars_name}{A list containing four vectors \code{confounders_internal}, \code{confounders_external}, \code{treatment_name}, and \code{outcome_name}. \code{confounders_internal} is a character vector containing the adjustment variables, which, along with \code{Estimator} and the corresponding \code{outcome_method} or \code{treatment_method} to correct for confounding; \code{confounders_external} is a character vector containing variables for weighting as to generalize estimates from \code{source.data} to \code{target.data}; \code{outcome_name} is a character vector of length one containing the variable name of outcome; \code{treatment_name} is a character vector of length one containing the variable name of treatment.}

\item{outcome_method}{A string specifying which model for outcome, treatment, and selection to use. Possible values are found using \code{names(getModelInfo())}. See \url{http://topepo.github.io/caret/train-models-by-tag.html}. A list of functions can also be passed for a custom model function. See \url{http://topepo.github.io/caret/using-your-own-model-in-train.html} for details.}

\item{treatment_method}{A string specifying which model for outcome, treatment, and selection to use. Possible values are found using \code{names(getModelInfo())}. See \url{http://topepo.github.io/caret/train-models-by-tag.html}. A list of functions can also be passed for a custom model function. See \url{http://topepo.github.io/caret/using-your-own-model-in-train.html} for details.}

\item{two_models}{An optional logical indicating whether potential outcomes should be modeled separately when \code{Estimator="DR"}. Default is FALSE.}

\item{outcome_formula}{An optional object of class \code{formula} describing the outcome model, treatment model, and selection model.}

\item{treatment_formula}{An optional object of class \code{formula} describing the outcome model, treatment model, and selection model.}

\item{...}{An optional arguments passed to \code{fit()} of each estimator object for model training and tuning. See \url{https://topepo.github.io/caret/model-training-and-tuning.html} for details.}
}
\value{
An object of class \code{\link{Estimator}}.
}
\description{
The function \code{Estimate} is used to estimate the average treatment effect obtained from \code{data}. If TEstimator is G_computation and outcome_method is BART, then make sure that treatment and outcome column is integer/numeric, no matter the outcome is binary or continuous.
}
